requirements.txt
# requirements.txt

# Core deep learning frameworks (PyTorch >=2.0 supports MPS on Apple Silicon)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLOv8 by Ultralytics
ultralytics>=8.0.0

# Computer Vision (webcam, image processing)
opencv-python>=4.7.0.72

# Audio capture
sounddevice>=0.4.6

# Numeric and image libraries
numpy>=1.22.0
pillow>=9.0.0